Nvim: make sure vanilla vim use `CursorLineNr` highlight
The implementation for `CursorLineNr` highlight group in vim and nvim
is different. In vim, you need to `set cursorline` first if you want to
use `CursorLineNr` highlight group. And also, you need to change
`cursorlineopt` from `both` (line and number) to `number` if you only
want to use `CursorLineNr` highlight and not `CursorLine` highlight.