Nvim: remap `Alt-~` (tilda) in insert mode
The default `Alt-~` in insert mode is `Esc-~`, basically it's escape
from insert mode and toggle uppercase/lowercase the previous character
which is counter intuitive for me. So I remap those to use `ctrl-o`
first before escape. With `ctrl-o` the cursor position didn't move so I
can toggle uppercase/lowercase for the character under the cursor.