Vim-sneak: limit the search only in visible line
Default behavior is search to the end of the file, which is not
effective because vim-sneak doesn't give label if it beyond the visible
line in current window. That's why limiting the search only in visible
line in current window makes more sense. If the pattern not in visible
line in current window, then mark it as `not found`. `lockvar` doesn't
have anything to do with it, I just forgot to uncomment it.