refactor(text editor): use `vim` instead of `nvim`
note: vim still spit out the output of external command to terminal but
found a work around for that, so it's not a problem anymore. what make
vim more sluggish seems like because the redraw in vim more noticeable
than in nvim, but after a while i don't really care. with that, all the
complaint about vim (kind of) solved.

plan: integrate vim config to nvim instead of using shell script to make
symbolic link.