vim: `lockmarks` instead of `k` command for move line mapping
note: if the length of above or below line is the same or longer than
the current line (the one that we move), then put the jumplist to the
same column as the current line. If the length of above or below line
is shorter than the current line (the one that we move), then put the
jumplist to the end of the line.