vim: `cn` and `cN` to substitute exactly the current word
note: previously `cn` and `cN` normal mode command substitute the word
or WORD under the cursor. now, `cn` and `cN` substitute exactly the word
or WORD under the cursor.

example(old): if the word under the cursor is `session`, it will substitute
the word `session`, `obsession`, `possession`, and so on.
example(new): if the word under the cursor is `session`, it will only
substitute the word `session`, and not `obsession`, `possession`, and so on.