vim: map `cd` and `cD` to substitute word and WORD in current line
note:
there's some downside about using `<>` in substitution as it's only
trying to substitute exactly the word or WORD inside of `<>`. so,
making a new mapping to not use `<>` and by default only in current
line because it's easier to add `%` at the beginning of command
instead of deleting it.