vim: first attempt join and split lines functionality
note:
the join function already works. if there's no count, it will behave
like the default `J` normal mode command. if there's count, it will
use the join function and we need to provide the direction. if we
want to join the entire paragraph, we need to use `1Jip` or `1Jap`.

the split function only works for the normal mode command. there's
still an issue when using the visual selection command. for some
reason, only the first occurrence that got substituted and not
respecting the indentation.