lynx: add lynx style sheet
note:
the default status bar text barely visible because of the background
color, so changed the background color to black. other than that,
everything still the default.

at the time of this commit, the default lynx style sheet can be found
on `/etc/lynx.lss` in arch linux (not sure about other distro).