vim(statusline): change condition for insert mode indicator
note:
there's a misunderstanding on my part. previously i thought that
`v:insertmode` will represent all the values that was in the help
menu, but it turns out that `v:insertmode` only represent the last
mode that use `InsertEnter` and `InsertChange` autocmd event.

so now, properly use the representation of insert, replace, and visual
replace mode in statusline mode indicator.