vim(statusline): start truncate line in filename component
note:
if the component become too long, rather than truncate the component
from the left side (which in this case is the mode indicator), start
the truncate at the filename.

so, the filename will be the first component to truncate if the filename
or other component too long. we can still see the filename by using
`ctrl-g` in normal mode, so it's not a problem. truncate other component
would be a problem or at least annoying to display manually.