vim(statusline): put square brackets in git branch name
note:
the components that appear only when a certain condition met using
the square brackets to enclose the component.

with that in mind, the only component that didn't have the square
brackets enclosing them is git branch component because i have no
idea how to do that before. now, make the git branch component use
square brackets like any other components that use conditional or
not appear by default.