vim(statusline): replace `autocmd` with `g:statusline_winid`
note:
as far as i know, in vim 8.1.1372 there's a new global variable that
has a window id of the active statusline window which is
`g:statusline_winid`.

if it's the same as the current active window id, then use the active
component of statusline. otherwise, use the non-active component of
statusline.

there's also a few references in the comment.