vim(statusline): fix mode indicator with `setwinvar()`
note:
now able to evaluate the change mode with `setwinvar()`.

the new problem is, the `{%` didn't seem available in vim 7. so this
method probably only works in vim 8 but this is still earlier than
global variable `g:statusline_winid` (if i remember correctly,
introduced in vim 8.1).

need to learn docker to test this out.