vim(ftplugin): change skip parameter for markdown mapping
note:
found out the issue why code block in markdown if code fenced syntax
highlight activated. it's because the syntax inside of code block will
use the syntax for specific filetype rather than `markdownCode` syntax.

for example:
if there's a shell script comment in code block, the syntax will use
`shComment` syntax highlight instead of `markdownCode`.

for now, i'll use the `shComment` instead of `markdownCode` for the
skip parameter in `search()`. need to make a regex for the skip
parameter.