vim(statusline): add percentage component
note:
the scenario when percentage in statusline can be useful is when
writing git commit message with `verbose` option. we don't have to
check if that's all the content for git commit message window
because percentage component will show `All` if that's all the
content for git commit message window.