vim(fzf): change keybindings for move halr-up/down preview
note:
`ctrl-b` and `ctrl-f` is useful to move to the left and right, so
change that to `ctrl-n` and `ctrl-p` for move half-page-up and
half-page-down.

the default `ctrl-n` and `ctrl-p` is move down and up, similar to
`ctrl-j` and `ctrl-k`. i already use `ctrl-j` and `ctrl-k` instead
of `ctrl-n` and `ctrl-p` so it's a waste not to use it for something
else.