vim(packages): not load lsp in gitcommit and vim filetype
note:
there's an issue that i didn't notice with previous implementation.
if we only load the lsp plugin on specific filetype, we can't really
load the lsp plugin when not open those specific filetype first.

here's an example:
let's say we open markdown file first, and then we open javascript
file. in those scenario we can't really activate the lsp because
the `vim-lsp-settings` didn't configure the `lsp_server` if not
entering vim (take a look at `VimEnter` autocmd event).

that's the only downside i've found with this particular lsp plugin.
fortunately, it's still fast as long as we didn't activate the lsp
server on startup (only load the plugin).