vim(statusline): indicator for combination space and tab indent
note:
if there's at least one tab character in a bunch of spaces indentation,
then this indicator will appear in statusline with the line number
of the tab character.

the flaw of this method is that, if the `expandtab` option set via
modeline and there's a tab character before adding the modeline, this
indicator won't be triggered. still not sure how to fix that at the
moment.