shell(aliasrc): limit the `sd` alias directory search
note:
there's quite a few times that i enter the wrong directory when
searching for all directory in the entire home directory.

so, now limit the directory to only two directory. the first is the
directory in home directory, and the second is the directory inside of
that directory. the directory after that is not included, use `ds`
alias instead.