vim(root): use `=` instead of special case for dot directory
note:
previously make special case for dot directory so that we didn't go
to that dot directory accidentally.

now, we need to explicitly give `=` (equal sign) to the pattern if we
want to change cwd to that pattern directory instead of the directory
above that.

for example, let's say we have `~/projects/.git` directory and we give
the pattern '.git', that will change our cwd to `~/projects`. if we
give the pattern `=.git`, that will change our cwd to `~/projects/.git`.

this is highly inspired by vim-rooter plugin, i just want to try using
`finddir()` and `findfile()` instead of `while` loop to search for the
pattern.