vim(editorconfig): fix error about `map()` argument in neovim
note:
previously, there's some error about the dictionary in `map()` argument
does not recognised as dictionary.

still not sure what happen, but i changed the logic a little bit so that
it returns dictionary in the first place instead of an array of
dictionary.

it's weird because in vanilla vim, the previous method works fine. hmm.