vim(editorconfig): only change local options value
note:
previously, changed the global and local value when found an
`.editorconfig` file. and then, i found out about `.editorconfig` file
in neoformat project which doesn't specify `indent_size` and `tab_width`
in general pattern (`[*]`) but provide `indent_size` and `tab_width` in
specific pattern (`[*.vim]`).

in that case, the value of shiftwidth or tabstop will stuck to the
latest file pattern that we access. that's not really a good idea. the
solution is that, only changed the local options value instead of both
global and local options value.