vim(editorconfig): add `try-catch` in `s:read_editorconfig()`
note:
there's this rare bug that when i try to open a file from filebeagle, it
will try to find `.editorconfig` file in current working directory
instead of root directory.

this is so rare that i'm not sure how to replicate this, for now let's
use `try-catch` and see if the error occur again.

*p.s: is it because of local buffer variable? but it only happens once
in a while? hmm.