vim(editorconfig): cache the editorconfig file path
note:
hopefully this method solve the random issue when opening the same file
and then use the full path of the current file with `.editorconfig` as
the parsing argument to editorconfig function.

example of full path of the current file with `.editorconfig`:
```sh
/home/username/project-directory/library/.editorconfig
```

while the actual path of `.editorconfig` file is something like this:
```sh
/home/username/project-directory/.editorconfig
```

which, of course, throw an error because we can't find the `.editorconfig`
file in the path of the current file.