vim(lsp): add diagnostics sign priority and specific hover mapping
note:
the diagnostics priority is set so that there's the error and warning
comes on top instead of the hint or other sign symbol. the default
priority for sign is 10.

for the hover stuff, rather than change the option of hover ui, use
specific plug mapping for hover float. no real reason, just
experimenting stuff.