vim(ftplugin): refactor markdown header movement
note:
instead of using `search()` over and over again, make a function and
reuse those function with other mapping.

there might be some drawback because we need to initialize function
first, but this method make it easier to maintain the code.

still need to improve this, especially the skip pattern.

todo:
if there's some comment that start with `#` inside a code block, ignore
it.