vim(ftplugin): skip markdowncode syntax highlight
note:
previously, if we have a comment that start with `#` and doesn't have
syntax highlight, it will use `markdownCode` syntax highlight and will
not be skipped by `markdown_header_movement()`.

now, i've figure out how to use backslash inside of double quote (or
maybe i've already figure that out but forgot?).

the issue for this approach is that, we can't go to the end of markdown
code block (```) because it use `markdownCode` syntax highlight.

haven't figure out how to exclude that in the `search()` skip pattern.