vim(ftplugin): fix markdown header movement for end of section
note:
previously we can't go to the end of heading section if it was a
markdown code block because we skip the `markdownCode` syntax highlight.

now, only skip `markdownCode` syntax highlight in non-end of heading
section movement, like `]]` or `[[`.

also, limit the search pattern of the end of heading section movement to
minimum 2 `#`. for example, we can go to the end of heading section
`## section 2` if the next section use `##` or more, if the next section
is `# section 1` then we can't go to the end of `## section 2`.

i mean, who would put the heading one in the middle of a file? if
there's such a case, i will change the regex again. for now, this fit my
need.