shell(gitrc): add git show `stat` in `gl` alias
note:
`stat` in `git show` is useful to check the filename of that we
committed.

there's a lot of `stat` option in git show (for more info check
`man git-show`), the one used here is `numstat`. why? so that there's
clear separation between the diff and the filename. i haven't take a
look at the customizability of `stat` in `git show`, so for now, this is
enough for my need.