vim(mappings): re-select newly copied line visual selection
note:
previously, when we copied using `vim-como` mapping, we can't visually
select the newly added line that we copied from another line.

`gv` re-select the previous visual selection, not the newly added line
from yanked text. so, we can't really use that.

fortunately, there's a mark to go to the first and last of the newly
added line from yanked or previously changed text, which is `'[` and
`']`. so, use that to select the newly added line from `vim-como`
mapping.