shell(gitrc): first attempt make interactive git add with fzf
note:
inspired by vim-fugitive, i want to add, remove, and check the diff of
the file in git working area inside of fzf as a selector.

this combine a few command, that is `git status`, `git add`, `git
reset`, and `git diff`. might add another command, who knows.

the downside for this selector that i have found is that, we can't
select the file that has space like `some horrible person make this.txt`
because fzf use a space as separation for each column. i don't really
want to support such a horrible person who make that kind of file name,
so expect this issue to be around for a long time.