vim(packadd): load vim-projectionist in all situation
note:
vim-projectionist is kind of hard to lazy load. even tho reloading all
the visible buffer "kind of" solve the problem, but it give another
problem and that is the position of the buffer is changing.

so yeah, load vim-projectionist all the time rather than getting a
headache about lazy load it.