shell(gitrc): first attempt replace `ggl` alias with function
note:
the mechanism is that, if there's no upstream reference for current
branch, do not pull.

if there's an upstream reference for current branch, check if the
upstream reference also exist on remote repo. if it exist, pull the
branch from remote repo. If it doesn't exist, do not pull and remove
upstream reference for current branch.

also, remove alias `gpl` for git pull because i rarely use it and forgot
about it.