shell(gitrc): fix remote name and branch name on `ggl` function
note:
when we want to get the remote name, we want the text before the first
slash and remove the first slash and the rest. on the other hand,
when we want to get the remote branch name, we don't need the text
before the first slash because that is the remote name.

previous implementation is incorrect because we get the text before the
last slash rather than the first slash for remote name and get the text
after the last slash for remote branch name.

here's an example, let's say we have branch `origin/test/anu/itu`.
the previous implementation will assign remote name as `origin/test/anu`
and remote branch name as `itu` which is incorrect because the remote
name is `origin` and remote branch name is `test/anu/itu` (this branch
name is possible).

the new implementation fix that. now we get the remote name from the
text before the first slash and the remote branch name from the text
after the first slash.