vim(mappings): use `ZA` keybinding for zooming keybinding
note:
we can close all the split window with ctrl-w ctrl-o, so using `ZA`,
which is easier to press to do the same can cause unexpected closing all
the split window.

rather than doing that, we might as well use it to zoom on specific
window if there are more than one windows.

also change tab variable from `zooming` to `zoom`.