vim(mappings): add skip expression for same indent movement logic
note:
now we can skip the same indentation before and after the current line
and go to the next occurrence.

for example, let's say we have this line and the cursor on `anu`:
```
  anu
  itu
    ini
  nganu
```

previously, the same indent movement will move to `itu` first and then
`nganu`. now, the same indent movement will move to `nganu`, skipping
the `itu`.

the downside of this approach is that, if there's no indentation, it
will also assume the blank lines (the lines without any character in it)
will be skipped.

for example, let's say we have this line and the cursor on `anu`:
```
anu

itu
ini
nganu

  ngitu
ngini
```

in the case above, the cursor from `anu` will go to `ngini` which
skipped `itu`, `ini`, and `nganu` simultaneously. this is not the
expected result.

maybe we can add another condition to not skip after or before the blank
lines with `prevnonblank()` and `nextnonblank()`? we'll see.