vim(mappings): fix same indent movement when no indentation
note:
previously, when there's no indentation, it will skip the line after the
blank line (the line without any character in it) which is not the
expected behavior.

now, we make sure only skip lines without indentation if previous or
next line is not blank line.

for example, let's say we have this line and the cursor on `anu`:
```
anu

itu
ini
nganu

  ngitu
ngini
```

previously, the cursor from `anu` will go to `ngini` which skipped
`itu`. now, the cursor from `anu` will go to `itu` and then `ngini`.