vim(mappings): use exact word under the cursor to highlight
note:
previously we highlight anything that has the word under the cursor.

for example, let's say we have word `anu` and `anuitu`. with the
previous pattern, if we put the cursor in `anu` we will highlight
both `anu` and `anu` in `anuitu`. now, we only highlight the word
`anu`.

should i use case sensitive too? let's observe first.