vim(grepprg): only use flags word regex on backspace mapping
note:
feels like we don't really need word boundaries or word regex in custom
grep and only need it in word and visual selection grep.

this is because if we use word regex, we can't do partial grep for the
custom keyword.

for example, if we want to check if there's any word that contain
`fold`, we won't find a result if there's only `foldclosed` and
`foldopened` but there's no `fold` (exactly fold, no more no less).