vim(motions): load vim-sneak from the start
note:
after recent update on my fork, i already move most of the function into
autoload files which means those function will not get loaded when
sourcing the plugin which increase startup time.

also, not using autoload function naming scheme in plugin files has
an effect such as vim did not try to look into autoload directory
(according to processes written in `verbosefile`) which could increase
startup time.

with that in mind, load `vim-sneak` should be slightly faster than
before and should not be a problem to load from the start.

another thing to add is that, we can't really access `vim-sneak` from
temporary normal mode from `ctrl-o` in insert mode. have not found a
workaround for that and using `mode()` as a conditions and `ctrl-o`
as action does not have any effect. so rather than adding another
unnecessary complexity, load `vim-sneak` from the start.