vim(lsp): use toggle lsp rather than automatic enable on filetype
note:
most of the time, i don't really use lsp. having the ability to turn on
or off the lsp server is better than enabling the lsp server from
beginning and take up a lot of resources just to view a file.

maybe i will add `autocmd` to enable lsp on mode changed? we'll see.