vim(mappings): use quickfix window for grep on cwd
note:
i need to be able to access the last grep result from current working
directory or custom path on every window rather than on specific window.
so using quickfix window feels better than location window which local
to current window, which means if i'm not on the window where i use
grep, i won't be able to access the list of the last grep results.