vim(mappings): store current window position in `t:zoom` variable
note:
feels like using zooming indicator just to store value 1 is redundant,
so store the command to restore to previous windows layout and then
execute those command if we want to restore to previous windows layout.

using `winrestcmd()` is better than using `wincmd =` which always trying
to set the window layout equal, and also does not work in quickfix
window because of `winfixheight` option in quickfix window.

Signed-off-by: bruhtus <hibruhtus@gmail.com>