vim(statusline): remove specific nvim format
note:
in neovim, we can place some info in the middle of statusline which kind
of cool until you realize that make the statusline not equal the more
info you have on your statusline.

it becomes a gimmick sooner rather than later. so remove that because
i don't miss it.