vim(qf): replace `quickr_preview.vim` plugin with simple function
note:
the problem with `quickr_preview.vim` plugin is that it uses cache for
the quickfix or location list, so when we use `colder` or `lolder` to
change to previous quickfix/location list, the cache won't be updated.

the result of out-of-date list cache is that:
- we can't use the preview.
- the preview display wrong file.
- the preview display wrong location in the same file.

so, make a simple function to replace the `quickr_preview.vim` plugin.

also, disable mapping to reload current command in quickfix window. this
mapping exist because i thought it would solve the `quickr_preview.vim`
plugin but it turns out the problem is in the list cache.

for reference, also add `vim-qf-preview` plugin for those who want to
use popup menu instead. but keep in mind that `vim-qf-preview` only
compatible with vim, not neovim.