vim(defaults): separate detect indent and set indent function
So that we can update the condition of current file on every save with
`BufWritePost` autocmd event. But, it only update the indentation info
about current buffer and not set indentation again. To do that, we need
to re-open the buffer so that the `BufRead` autocmd event triggered.

Not changing the indentation setting on save is intentional, to prevent
inconsistent indentation. Set it once, and if we want to reset the
settings, re-open the buffer using `:e` command.

Using `search()` function on every save should not burden vim too much,
so it should be fine (?). We'll see.