vim(defaults): use separate function for editorconfig indentation
note:
the previous mechanism still searching for editorconfig file because
the default when `b:editorconfig_file` empty is search the editorconfig
file which means the condition to not search in specific buffer is
useless.

forgot when i changed the condition from `b:editorconfig_file` exist to
`b:editorconfig_path_cache` exist, but because of that changes, the
mechanism become quite different from the initial plan.

now, use separate function to make controlling when to use editorconfig
easier.