vim(statusline): move most indicator to the right side
note:
the indicator that get moved to the right side are:
- indent
- zoom
- session
- lsp

other than that, move the percentage to the left side so that the
filename is not in the left corner.

also, shorten the git branch name when the window is small from 10 to 6
character to give more room for other indicator in a small window.