zsh(zshrc): disable command syntax highlight
Feels like it's more readable to compose a command without a color,
especially on tty where the yellow color is hard to see.

`zsh-syntax-highlighting` use yellow color for all the string and that
makes look at the string on tty harder. We could change that, but it
requires a lot of env variable. So for now, turn off syntax highlight
and see how it goes.