i3(config): trying specific workspace workflow
The concept is using specific workspace to open a specific app.

For example, let's say we want to open all of our web browser in
workspace 2 and we want to open our terminal in workspace 10. If we want
to go to our web browser, we need to open workspace 2 and to go back to
our terminal, we need to open workspace 10.

This workflow let us minimize the `alt-tab` nightmare of cycling a bunch
of program and hopefully remove the overhead of "where did i open this
program".

I'm still getting used to it, so let's see how it goes.