vim(defaults): always print filename when grep one file
By default, if we only search in one file, `grep` and `ripgrep` does not
show the filename which means we can't use it with quickfix window if we
only search in one file.

To make `grep` and `ripgrep` show filename even when searching in one
file, we need to use `-H` or `--with-filename` flag.

Also edit typo aplhabet to alphabet.