vim(lsp): use default nmap ctrl-] to go to definition
Because we already set `tagfunc` to `vim-lsp`, we can use default
normal mode map `ctrl-]` to go to definition.

There's some minor different between using tag and the default go to
definition mapping, is that we go to the first non-whitespace character
when using tag instead of the exact place when using go to definition
mapping.

Let's see how it goes.