vim(ftplugin-qf): set `eventignore` before open preview window
Just realized that the autocmd event still triggered when we open the
preview window. That is because we open the preview window first rather
than setting the `eventignore` first, which means `eventignore` option
have no effect on preview window.

Now, setting `eventignore` option before we open the preview window.

Also make sure that the `buftype` conditional is case-sensitive.