vim(fugitive_log): use full path when trace line number log
In case our current working directory is not a git repo and we have a
git repo inside it. For example, opening vim plugin in `pack` directory
from `~/.vim`. If we use relative path with that scenario, the result
would be something like this:
```sh
git log -L 1,2:pack/minpac/opt/.../plugin/plugin.vim
```
which result in an error that `pathspec` does not exist.

To solve that, we need to give the full path of the current file so that
we're not limited to current working directory.