vim(defaults): limit editorconfig search to home directory
Previously i use `findfile()` which search the file until root
directory, and i can't figure out how to use the stop directory.

Now, to make things simple, we use while loop instead so that we can
control when to stop. Don't really plan to using editorconfig on root
directory, so this should be fine.

Also, improve the condition to not search the editorconfig file and
refactor the function name to make it different from variable name.