vim(formatter): use nmap `Q` to manually format current buffer
Rather than set global variable to disable formatter, switch to manual
formatting instead.

The downside of manually format the buffer is that, we might forget to
format our buffer using the formatter. But, the upside is that, we don't
need to disable the formatter if we don't want to and we can format the
buffer only when we need.

Let's see how it goes.

Also, remove `!` from `unlet` because we already checking if the local
buffer variable exist or not.