vim(insanity): change StatusLineNC and TabLine fg to yellow
First, change the yellow color to `100` or `#878700` so that it does not
appear dark red on tty and still resembles dark yellow on terminal
emulator.

And then, use the new yellow color for `StatusLineNC` and `TabLine`
foreground so that we can easily distinguish the non-active window and
tab, and also read the content of those non-active window and tab (such
as filename).

Because we can already see the filename on the non-active window, we
don't need to specifically use `StatusLine` highlight for filename on
our statusline.