tmux(conf): use default length and prefix `<` on status left
Feels like we don't really need 85 characters for our status left
component which only for session name. So using the default status left
length, which is 10 characters should be enough.

Also, truncate the session name if it's more than 6 characters to make
it fit into the 10 characters maximum length (2 characters for square
bracket, 1 character for space to separate the first pane title).