vim(lsp): use the same mapping for signature help
Previously we use the `<Cmd>` mapping for signature help if we have
patch 8.2.1978 or using `ctrl-\ ctrl-o` if we don't have the patch. Now,
only use `ctrl-\ ctrl-o` imap instead of `<Cmd>`.

The previous problem is that we use imap `ctrl-o` to access normal mode
command and then it move the cursor when we use it at the end of the
line, now we use imap `ctrl-\ ctrl-o` which is like imap `ctrl-o` but
does not move the cursor even on the end of the line.