polybar(config): use interface-type for network module
To make polybar more flexible, trying to use `interface-type` and let
polybar detect the network interface automatcally.

This approach might have a problem if we have multiple network
interfaces, because polybar will use the first network interface found
which might or might not be the correct interface.

Also, this options only available in version 3.6.0 or newer.