i3(config): add focused_tab_title colors
focused_tab_title colors use the same colors as focused colors, except
the indicator and child_border colors.

The indicator and child_border colors of focused_tab_title use the same
colors as focused_inactive indicator and child_border color.

This is to prevent confusion with the inactive client in tabbed mode.