zsh(zshrc): display menu completion when list not fit on screen
Previously when there's a lot of completion list and not enough space in
the screen, zsh would print out all the completion list and we need to
scroll up from our terminal or tmux.

With the `yes` or any of the true values (`yes`, `true`, `on`, `1`), zsh
will use menu completion. And if we use `yes=long`, we can limit the
menu completion to show only when the completion list does not fit on
the screen.

Why do i want to only show menu completion when the completion list does
not fit the screen? Because menu completion will automatically select
the first item on the completion list, unless we can change this
behavior with `zsh/complist` which i haven't look at.