feat(zshrc): use menu completion interactive
After more reading `man zshcompsys`, we can insert `interactive` option
so that we can enter `interactive mode` which let us type the word
without exit the menu completion.

Keep in mind that `interactive mode` does not use fuzzy search. For
example, if we want to complete `.zshrc` file and we only type `rc`, it
won't find the `.zshrc` file. We must type at least `.z` to make the
completion `.zshrc` appear.

Also, because we use interactive mode from the start, it will conflict
with the vim-like keybinding for menu completion. For example, let's say
we want to type `l` in the interactive mode. If we still have the
vim-like keybinding in menu completion, we will exit the interactive mode
and go to left or the next completion word.

At the moment, have no idea how to prevent the vim-like keybinding
intervene with interactive mode.