zsh(zshrc): remap tab to expand or complete from prefix
Previously when we expand or complete the word from the word under the
cursor.

For example, let's say we first type `ls` and now we want to add `sudo`
in front of it. Previously when we type `suls` and the cursor is on the
`l` word, we can't trigger the expand or completion. Now, when we type
`suls` and the cursor is on the `l` word, we can trigger the expand or
completion and choose `sudo` which result in `sudo ls`.