vim(defaults): add switchbuf uselast option
When we jump to the line from quickfix and we have more than one window
in split, the window used to access the file is not consistent.

`uselast` option in `switchbuf` make sure we use the last window we used
when we jump to the line or file from quickfix. This is useful to
prevent confusion where the file will be opened.