shell(gitrc): add safety guard when ggl option need argument
Trying to differentiate the error between invalid option and an option
that need argument.

If we didn't add `:) ... ;;`, the error will go to the `?) ... ;;`
instead and will print out usage with the correct option, which
sometimes lead to confusion.

Also, remove the `h` option in favor of `?` option, which catch every
option except the specified one.