shell(dotfiles-manager): replace dotbare with sdfm
Trying to implement my own dotfiles manager so that i don't need to
download a zsh plugin just to manage my own dotfiles repo.

I named it sdfm, which stand for simple dot file manager, because i'm
lazy at naming stuff and trying to prevent any conflict with existing
command by using 4 character instead of 3.

If you are wondering, we do we use shell function for the `git add`
functionality?

It's because i can't seem to make the filename completion works with the
shell alias in zsh, that's why i use shell function and `compdef` to add
filename completion for `git add` functionality.