qutebrowser(config): limit switch tab to T keybinding
Trying to make qutebrowser like my current vim UI, which only has status
bar.

To switch between buffer on my vim? Use fzf which appear in the bottom,
similar to the `T` keybinding which appear at the bottom too. The
difference might be in the fuzzy search, but for searching url, that
might not be a problem. We'll see how it goes.