qutebrowser(config): remap escape normal mode to add stop loading
The default `escape` normal mode keybinding is
`clear-keychain ;; search ;; fullscreen --leave` which i have no idea
what `fullscreen --leave` means. From the description of `fullscreen`
command, it's for toggling fullscreen which means it's for enter and
exit fullscreen. So, what's the point of `fullscreen --leave`? Is it a
legacy stuff? Probably.

Anyway replace `fullscreen --leave` to `stop` for stoping the loading
webpage, similar to default `ctrl-s` normal mode.