i3(config): try using title bar
The problem when not using title bar is that finding the current
focused window harder.

For example, let's say i'm opening a web browser that took awhile to
open, and in the meantime i interact with my scratchpad of terminal
emulator. By the time the web browser open, i lost focus of the
scratchpad and become confused which one i'm currently focused on.

So let's try using the title bar and see if it can lessen the confusion
about focused window.