shell(sdfm): prevent use partial init after full init
The problem with previous implementation of partial dotfiles init is
that we can do partial init even after the full init, which is not
intended.

If we can do partial init after full init, sooner or later we will be
confused with the current state of our local dotfiles repo.

Even though `sdfm` use `git checkout` as partial init implementation,
that does not mean we should use partial init as replacement for
checkout command, which i assume some user will do.