zsh(zshrc): remove hist_verify option
Just found out that we can use history expansion `!$` to change
directory into the cloned git repo.

For example, we clone this dotfiles repo with command:
```sh
git clone https://github.com/bruhtus/dotfiles
```
and then we want to change directory into the cloned repo, we can do
this:
```sh
cd !$:t
```

Please keep in mind that `!$` is not POSIX compliant, so it might not
work in some shell.

If you want something that will work in POSIX compliant shell, you can
use parameter expansion instead, like this:
```sh
cd ${_##*/}
```

Now, the problem with `hist_verify` option is that when we use any
history expansion and press enter, it will print out the content of the
history first instead of execute the specified history.

If we use the example above with `hist_verify`, when we enter this
command:
```sh
cd !$:t
```
it will print out:
```sh
cd dotfiles
```
instead of change directory to `dotfiles`.

With `zsh`, we can use `tab` to expand the history if we're not sure
anyway. So remove `hist_verify` option and see how it goes.