shell(gitrc): use fzf reverse-list layout for `gurm` alias
The downside of default fzf layout is that, when we use multiple
selection and select the current line, the cursor or indicator will go
down automatically.

Let's say there are 3 items, the first item is the one on the bottom.
We select the first item, and then we select the second item. After we
select the second item, the cursor will go down to the first item. So if
we want to select the third item, which is on the top, we need to go up
twice just to select the third item.

`reverse-list` layout solve those inconvenience.