tmux(conf): only show window flag as current window format
At first i thought that by removing the index in current window
indicator can reduce the confusion when switching window, and it turns
out that i still confused which one the current window and which one the
non last active window.

So trying to simplify the current window indicator by only showing
window flag and see how it goes.