vim(mappings): remap zoom window to <leader>z
Make zooming window mapping similar to how tmux do it.

The reason i didn't do this before is because of preview mapping in
quickfix window. Previously when we want to preview the file in quickfix
window, we use `<Space>` which is basically my leader key. Now, i
changed that preview mapping to `p`, which means i can use my leader key
in quickfix window.