tmux(conf): skip empty prompt from search beginning of command
The point of searching the beginning of previous command is to make
reading the output of those command from beginning easier.

So, by omitting the empty prompt, which is a prompt that does not have
a command in its line, we can speed up searching previous command that
we want to see the output.