qutebrowser(config): add insert mode map for change to passthrough
The benefit of passthrough mode is that, we can use all the default
keybindings such as `ctrl-a` to select all text. In insert mode, i
bind some key so that i have the same keybindings as readline
keybindings.

The keybinding i use to change from insert mode to passthrough mode is
`shift-esc`, similar keybinding to leave the passthrough mode.