i3(config): open terminal in ws 3, slack and telegram in ws 7
Trying to change my workflow a little bit. Opening the workspace 10
with keybinding `alt-0` has the potential to miss press the
keybinding `alt-minus` which will lock the screen.

So, to prevent that, let's try using workspace 3 for terminal stuff
and move the chat app such as slack and telegram to workspace 7 and
see how it goes.